==> 11-filter-cs-ai-lg.ipynb <==
import pandas as pd

# Download ist zu groß
df = pd.read_parquet("arxiv-2025-02.parquet")

df.dtypes

llm_mask = df["categories"].apply(lambda x: ('cs.LG' in x) or ("cs.CL" in x))

ai = df[llm_mask].copy()

len(ai)

ai[ai["year"] >= '2023'].to_parquet("arxiv-2023-llm.parquet", compression="zstd")
