==> 21-embed.ipynb <==
from tqdm.auto import tqdm
from sentence_transformers import SentenceTransformer

Modell ermitteln: https://huggingface.co/spaces/mteb/leaderboard

#model_name = "jinaai/jina-embeddings-v3"                 # am schnellsten
#model_name = "intfloat/multilingual-e5-large-instruct"   # etwas langsamer
model_name = "Snowflake/snowflake-arctic-embed-l-v2.0"    # ca. genau so schnell
#model_name = "infly/inf-retriever-v1-1.5b"               # etwa 3x langsamer

model = SentenceTransformer(model_name, trust_remote_code=True).cuda()

import json
import lzma
with lzma.open("llm-abstract-sentences.json.xz", "rt") as f:
    es = json.loads(f.read())

len(es)

%%time
# Test zur Geschwindigkeitsmessung
embeddings = model.encode([e["title"] + ": " + e["text"] for e in es[0:1000]])

embeddings.shape

# alle Embeddings ausrechnen
embeddings = model.encode([e["title"] + ": " + e["text"] for e in es])

import numpy as np
with open("llm-abstract-sentences-saev2.npy", "wb") as f:
    np.save(f, embeddings)

qe = model.encode("How long do I train an LLM?", prompt_name="query")

similarities = model.similarity(qe, embeddings).flatten().numpy()

for i in similarities.argsort()[::-1][0:5]:
    print(es[i])
